Reduce the number of test permutations in AOCO compression tests.
Some tests in the AOCO compression TINC test suite were run with 1MB and
2MB block sizes. That seems excessive. There is no magic boundary, like
e.g. when going from 1 GB to 2 GB, between 1 MB and 2 MB, so it's hard to
see how anything would behave differently in any interesting way between
those two values. This test suite runs for a very long time, so remove
the 1 MB permutations from those cases, to reduce the runtime.

(The test files have 1G and 2G in the names, but that is misleading; these
tests vary the blocksize option to CREATE TABLE command, between 1MB and
2MB. That's megabytes, not gigabytes. That ought to be fixed, but not in
this patch)