Move more tests for fixedwidth formatter from legacy cdbfast test suite.
All the readable tests are moved to a new regression suite in
contrib/formatter_fixedwidth, next to the source. Writable tests, however
are added to gpfdist's 'custom_format' test, which already contained one
writable test like this (query07, in the old test suite).

Some minor changes to the original tests:

* Use file:// urls instead of gpfdist:// urls, for the readable tests.
  That's simpler, as you don't need a running gpfdist server.

* Left out a couple of redundant tests. Queries 6 and 7 were already in
  gpfdist's 'custom_format' test suite.

I kept the original numbering, so that you can relate to the old test
suite. That will become obsolete as soon as this is committed, but might
be useful for reviewing.

Signed-off-by: Adam Lee <ali@pivotal.io>