Fix the crash / error that happens when we have bitmap or btree or both in qual clause.
Revert(except test) "[#129402309] Fix double free when using stream bitmaps with hash bitmaps."

This is ported from
https://github.com/greenplum-db/gpdb/commit/3d69ae41834d4de0e1dfbbcdfbba237f207ea67b.

This reverts commit b5d7b30fd36787a0cb3a7d28a2dadc9e259deda2.

`HashStreamOpaque` will never own the `HashBitmap` and hence there is no
need of variable `free` in `HashStreamOpaque` to denote if it owns the
`HashBitmap` or not. Hence we don't need `tbm_create_stream_node_ref`
API at all. Given that `HashStreamOpaque` is not owning the `HashBitMap`
it should never try to access the `HashBitMap` during the free of
`HashStreamOpaque`. There is no guarantee that `HashBitmap` is valid at
that time. `ExecEndBitmapIndexScan` might have freed the `HashBitmap`
that is referenced in `HashStreamOpaque`.

Fix HashStreamOpaque from bad access / double free'ing HashBitMap.

This is ported from
https://github.com/greenplum-db/gpdb/commit/57b9aeb38f76f55e6f09c1281bdb30eb636f9886

`HashBitMap` is owned by `BitmapIndexScanState`(`node->bitmap`).
`BitmapIndexScanState` will free the `HashBitMap` when `ExecEnd` is
called. So it is not the responsibility of `HashStreamOpaque` to free it
which just has the reference to it.
Also, it should never try to access
the `HashBitMap` when `HashStreamOpaque` gets freed. `HashBitMap` might
have been freed even before the `HashStreamOpaque` gets freed.

Similar to optimizer plan, make executor code to free the bitmap for planner plan.

This is ported from
https://github.com/greenplum-db/gpdb/commit/08112e2fa05cc8bcc91cb06b82a14bae5fcc4d4c

Similar to the optimizer, in planner with this
change`BitmapIndexScanState` is responsible for freeing the bitmaps
during `ExecEnd` or `ExecRescan`. The parent node doesn't need to worry
about freeing the bitmaps.

This also fix the following error for planner plan(bitmapscan_ao.sql).
```
select count(1) from bmcrash b1, bmcrash b2 where b1.bitmap_col =
b2.bitmap_col or b1.bitmap_col = '999' and b1.btree_col1 = 'abcdefg999';
ERROR:  non hash bitmap (nbtree.c:572)  (seg1 slice2 127.0.0.1:25433
pid=83873)
```

Introduced the new api to add streambitmap from another streambitmap.

This is ported from
https://github.com/greenplum-db/gpdb/commit/11f815cbdc9cb980d8c9c8326d26b0b3dcb81a22.