Changed function used to test CAST filtering.
The Behave tests written to test CAST filtering on master used a
"castToInt" function that performed a text-to-int cast.  Because the
built-in implicit casts differ between 4.3 and 5, such a cast already
exists in 4.3, so it has been replaced with a boolean-to-text cast.
Neither cast is actually used anywhere outside of tests, so the CAST
filtering functionality in gprestore_filter.py is unchanged.