Do not eager free nested subplans when a subplan finishes.
When a subplan finishes, it eager frees any nested subplans. However, if
the subplan is in a qual list (of a SeqScan e.g), it may be executed
multiple times; and if the nested subplan has a Material node, all the
materialized data is abandoned. This is a problem if the Material is on
top of an operator that can no longer regenerate the data (e.g Motion).

This issue was fixed in 5.0 along with the 8.3 merge. This commit brings
relevant changes into 4.3 by only eager freeing nested subplans when the
operator containing the top most subplan is finished.