Coverity fixes for walreceiver.c
CID 129947: Out-of-bounds write (OVERRUN)
overrun-local: Overrunning array buf of 80 bytes at byte offset 80 using index
nbytes (which evaluates to 80).
The above defect was found in `WalRcvUsr2Handler()` which is in gpdb only and
not in upstream.

CID 129271, 129270, 129269, 129268 : Function address comparison (BAD_COMPARE)
func_conv: This implicit conversion to a function pointer is suspicious:
The above defect was found in `WalReceiverMain()`. The comparison with NULL was
being performed against a function name and not a function pointer or variable
which is clearly incorrect. Since we do static linking of the functions, the
check was not necessary in the first place. Fixed the defect by removing the
check.