Fixup subplans referring to same plan_id
Before parallelization on nodes in cdbparallelize if there are any subplan
nodes in the plan which refer to the same plan_id, parallelization step
breaks as a node must be processed only once.
This patch fixes the issue by generating a new subplan node in glob
subplans, and updating the plan_id of the subplan to refer to the newly
created node.