Inline simple CTEs as a first stage planning [#144635043]
We convert simple CTEs to subqueries if the CTE does not contain another
CTE, it is not corralated and it is not refered more than once.

This helps us to avoid creating plans like below:
```
 Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..56023.00 rows=1000 width=8)
   ->  Seq Scan on foo  (cost=0.00..56023.00 rows=334 width=8)
         Filter: (subplan)
         SubPlan 1
           ->  Subquery Scan x  (cost=0.00..28.00 rows=1 width=0)
                 ->  Result  (cost=28.00..28.01 rows=1 width=8)
                       Filter: public.foo.a = $0
                       ->  Materialize  (cost=28.00..28.01 rows=1 width=8)
                             ->  Broadcast Motion 3:3  (slice1; segments: 3)  (cost=0.00..28.00 rows=1 width=8)
                                   ->  Seq Scan on foo  (cost=0.00..28.00 rows=1 width=8)
```

Because of the explicit CTE check in `is_simple_subquery`, the subquery
do not gets pulled up to have an inexpensive `Hash EXISTS Join` plan.

This caused some problems in customer queries and this commit helps to
generate better plans if the CTEs are simple.

We can disable this feature by setting `gp_inline_simple_cte` to off.
It is enabled by default.

Signed-off-by: Bhunvesh Chaudhary <bchaudhary@pivotal.io>