Fix dumping of "COUNT(*) FILTER (...)" expressions.
Without this fix, the FILTER expression would be left out of the deparsed
DDL of a view. Now it gets dumped as the CASE - WHEN expression that we
tranform the FILTER to at parse analysis. Ideally, we would dump it using
the original FILTER syntax, but that would be a much bigger patch. We'll
get that when we merge the upstream FILTER implementation, in PostgreSQL
9.4.

Fixes github issue #1854, reported by @water32.

(cherry picked from commit 54d6b71b57fc1a41bc5bed9316db0c453d3963fa)