gpperfmon: Fix incorrect data saved into queries_history table
sigar_proc_cpu_get is the function which takes a pid
and goes to the cpu and gets the cpu_elapsed value for that pid at a
given moment.

get_pid_metrics saves the information into a hash table within the
gpsmon process. That pid only gets wiped out of the hashtable when
gpmmon requests a dump command to gpsmon.

gpsmon process sends tcp packets with information about cpu_elapsed
(amount of cpu cycles spent on a given
slice/process). Doing a 'dump' clears out the gpsmon hashtables.

This could lead to scenarios where a query has ended,
the process has died, and the sigar_proc_cpu_get function can't find the pid
so it puts 0 into the struct which is updating the hashtable so the
hashtable has a 0 for cpu_elapsed and then this is what is in
queries_tail and so queries_history ends up having a 0 for cpu_elapsed
since it was the last entry in queries_tail

Fix: Ensure that we validate the functions that request pid metrics
from libsigar and log the issues if they occur.

Signed-off-by: Tushar Dadlani <tdadlani@pivotal.io>