TINC test to detect a bug in filerep resync logic.
Filerep resync works by obtaining blocks changed since a mirror went down from
changetracking (CT) log.  The changed blocks are obtained in fixed sized
batches.  Blocks of the same relation are ordered by block number.  The bug
occurs when a higher numbered block of a relation is changed such that it has
lower LSN as compared to lower numbered blocks.  And the higher numbered blocks
is not included in the first batch of changed blocks for this relation.  Such
blocks miss being resynchronized to mirror due to incorret filter based on
previously obtained changed blocks' LSN.  That means the mirror is eventually
declared in-sync with primary but some changed blocks remain only on the
primary.  This loss in data manifests only when the mirror takes over as
primary, upon rebalance or the primary going down.