gppkg: sync up a migrated gppkg to all segments
Currently, doing a gppkg --migrate will migrate the gppkgs from one
gphome to another. However, this is not done on the segment hosts as
well. So when a user tries to run a gppkg installed sql function, it
will break.

Fix: sync up the migrated packages from the master gphome to all segments

-Add tests to cover MigratePackages
-Add more unit test coverage for gppkg

Signed-off-by: Nadeem Ghani <nghani@pivotal.io>