Determine path to R from the RPM of R in plr gppkg
The path to the correct R package in the morgan stanley build script
should not be using a hardcoded version. In place of hardcoding, the
version is determined by looking at the version field of the RPM that is
included with the plr gppkg file.

This doesn't protect against changes of path structure in the R package,
it only protects against version changes.

Signed-off-by: Kris Macoskey <kmacoskey@pivotal.io>