Remove inaccurate assertions
When we DELETE or UPDATE an inherited table, some tables in the
inheritance tree might need an explicit Motion node to bring the targeted
tuples back to the segment where they reside, and some might not. The code
to build the plan handled that correctly, but this assertion incorrectly
assumed that it's all or nothing.

Moreover, the subtrees under APPEND may not even be rooted at MOTION's
at all: consider an inner join between a (root) partition table and
another table, the natural plan would be a (HASH) JOIN (on the outer
side) on top of an APPEND of many leaf partitions, and on the inner side
a HASH node on the "other" table. But an equivalent plan could be an
append of many inner JOINs, the inner side of each of which being the
HASH subtree on the "other" table, but the outer side being a SCAN of
a leaf partition. In the latter scenario, a child of APPEND is not even
a MOTION, let alone an EXPLICIT MOTION.

This commit removes the assertion, as it doesn't seem very useful in the
first place.  The code that inserted the Motion nodes is just above the
assertion, and the assertion was basically just testing the same thing
that the code just did, and not some general invariant that should
always hold.

Fixes issue greenplum-db/gpdb#332

Original author is Heikki Linnakangas.
Cherry-picked from
greenplum-db/gpdb@04fe0202c59721826d1eda2b19d73e5572893fcb with some
wordsmith-ing on commit message.

(cherry picked from commit 04fe0202c59721826d1eda2b19d73e5572893fcb)