Reduce the number of join orders considered by GPORCA
Previously we were setting the value of
`optimizer_join_arity_for_associativity_commutativity` to very large
number and so GPORCA would spend a very long time evaluating all possible n_way_join
combinations to come up with the cheapest join tree to use in the final plan it produced.

We are reducing this value to `7` as it does not prove to be beneficial
to spend time and resources to evaluate any more than 7_way_joins in
trying to find the cheapest join tree.