Refactor MS build to enable testing (#348)
In order to facilitate the reuse of teams tests, we have to "patch" the
installer and add in file and over write versions of other files that
those tests need.  We are opting to build a pipeline that exposes both
the patched installer and the "shippable" installer to each test. It is
important that tests don't pollute or modify the shippable artifact.
This was done so that we can guarantee the version of the arifact that
we ship matches the one that we tested.  There is risk in this approach
that we feel is warranted given that MS will soon be changing its method
for consuming artifacts from us.

Signed-off-by: Alexandra Wang <lewang@pivotal.io>
Signed-off-by: Kris Macoskey <kmacoskey@pivotal.io>
Signed-off-by: Jim Doty <jdoty@pivotal.io>
Signed-off-by: Divya Bhargov <dbhargov@pivotal.io>