Changes necessary for concourse 3 and compatible with 2.7.3
Migrations efforts for concourse 3 that are backwards compatible with
concourse 2.7.3

Includes using the image_resource in external task yamls for pipeline
jobs that use a sles based docker image. Currently unknown why the sles
image in particular causes image_resource issues.

As well, a generic task script for compiling GPDB was modified to do a
chown slightly differently, and with a retry, but it does not change the
expected end result in any way. This change was necessary to handle a
peculiar bug in concourse 3 such that chown fails in unexpected ways.

Signed-off-by: Divya Bhargov <dbhargov@pivotal.io>
Signed-off-by: Jim Doty <jdoty@pivotal.io>